Refactor components and systems
- Move components and systems to their own files grouping them in folder by types
- Move game logic specific components to sandbox
- entt use uuid for entity id now
- avoid creating prefab if already exists (use update for that)
- add create, update, update all prefabs
- update all components to send string, vec, glm vec3 and some methods through the engine to send values by &&
- remove components initialization logic from entity factory to onComponentAdded in each component file, texture in utils scene for now due to circular dependency bug
- render now checks if the shader program exists
- throw error if prefab already exists
- new methods in entity class to return scene and to update a component from the entity
- refactor assign and assign values in the same method, new method trigger event
- trigger event call scene onComponentAdded, remove redundant onComponentAdded in entity class
- add unsupported onComponentAdded exception in the default template
- optimize entity parameter in scene methods
- prefabs creations done in game now instead of scene class initilization to allow user defined prefabs with custom components and onComponentAdded methods
- replace const& string for std_view
- fix const& string for move semantic when we store the string
- use move and emplace back when possible
- guards preventing invalid values in cTime component, default values for avoiding boilerplate
- add cast component and print for each component so we can print them in runtime without knowing the type
- new method onComponentCloned and logic
- remove skyColor, use instead fogColor
- start using monostate for global uniform variables (light, sky)
- update to entt new version replacing deprecated methods emplace/push, get/value
- fix ligthColor being lost due to wrong type
- missing CCollider trigger
- rename efactory to entityFactory
- fix prefabs not being deleted due to being filter out
- disable specular from shader until fixed
- delete unused loadTextures method from entity factory

+++ potatocraft/potatoengine/utils/uuid.h
@@ -5,16 +5,16 @@ namespace potatoengine {
 
 static std::random_device rd;
 static std::mt19937_64 gen(rd());
-static std::uniform_int_distribution<uint64_t> dis;
+static std::uniform_int_distribution<uint32_t> dis;
 
 class UUID {
    public:
     UUID() : m_uuid(dis(gen)) {}
-    UUID(uint64_t uuid) : m_uuid(uuid) {}
+    UUID(uint32_t uuid) : m_uuid(uuid) {}
 
-    operator uint64_t() const noexcept { return m_uuid; }
+    operator uint32_t() const noexcept { return m_uuid; }
 
    private:
-    uint64_t m_uuid{};
+    uint32_t m_uuid{};
 };
 }
